chussa login  haleko

+++ admin/pages/delete/user.php
@@ -1,11 +1,15 @@
 <?php
-include "../../config/config.php";
-$id = mysqli_real_escape_string($conn, $_GET['id']);
 
-$sql = "delete from users where id='$id'";
-if ($conn->query($sql)) {
-    $_SESSION['branch_deleted'] = "successful";
-} else {
-    $_SESSION['branch_deleted'] = "error";
+if (isset($_GET['id'])) {
+
+    include "../../config/config.php";
+    $id = mysqli_real_escape_string($conn, $_GET['id']);
+
+    $sql = "delete from users where id='$id'";
+    if ($conn->query($sql)) {
+        $_SESSION['branch_deleted'] = "successful";
+    } else {
+        $_SESSION['branch_deleted'] = "error";
+    }
 }
 header("Location:" . url . "users.php");
